Adopt a new syntax for specifying configuration for subcommand
Previously to specify configuration for a subcommand, you use:

    [SUBCMDNAME]
    param=val

and:

    [SUBCMDNAME profile=PROFILENAME]
    param=val

We now use:

    [subcommand=SUBCMDNAME]
    param=val

and:

    [subcommand=SUBCMDNAME profile=PROFILENAME]
    param=val
    ; or
    [profile=PROFILENAME subcommand=SUBCMDNAME]

Basically, we now regard the section name as a whitespace-separated list of
key=value pairs. The new syntax is clearer and more generic. Application can add
their own key=value as needed, for example a backup utility can use:

    [backup=jono]
    source=/home/jono
    target=/backup/jono

    [backup=ujang]
    source=/home/ujang
    target=/backup/ujang